+++ install.php
@@ -10,14 +10,16 @@ rex_sql_table::get(rex::getTable('pagestats_data'))
 
 rex_sql_table::get(rex::getTable('pagestats_visits_per_day'))
     ->ensureColumn(new rex_sql_column('date', 'date'))
+    ->ensureColumn(new rex_sql_column('domain', 'varchar(255)'))
     ->ensureColumn(new rex_sql_column('count', 'int'))
-    ->setPrimaryKey(['date'])
+    ->setPrimaryKey(['date', 'domain'])
     ->ensure();
 
 rex_sql_table::get(rex::getTable('pagestats_visitors_per_day'))
     ->ensureColumn(new rex_sql_column('date', 'date'))
+    ->ensureColumn(new rex_sql_column('domain', 'varchar(255)'))
     ->ensureColumn(new rex_sql_column('count', 'int'))
-    ->setPrimaryKey(['date'])
+    ->setPrimaryKey(['date', 'domain'])
     ->ensure();
 
 rex_sql_table::get(rex::getTable('pagestats_visits_per_url'))