+++ Changelog.md
@@ -1,5 +1,20 @@
 # Changelog 
 
+
+## [xx] - xx.xx.xxxx
+
+### Added
+
+### Changed
+- sortable lists now sort dates correctly
+
+### Removed
+
+### Vendor Updates
+
+
+
+
 ## [2.0.0-beta.12] - 12.12.2021
 
 ### Added
@@ -14,6 +29,7 @@
 ### Notes
 Dieses Update beinhaltet auch die Änderungen aus den Betas von 2.0.0-beta.11 und 2.0.0-beta.10.
 In diesesn war allerdings die update.php fehlerhaft, weswegen die Referer-Daten nicht korrekt migriert wurden.
+
 
 
 ## [2.0.0-beta.11] - 11.12.2021